chore(poupe-vue): components: move onSlot() to a dedicated file
